Split SearchServiceClient into separate index and indexer clients
Fixes #12151
