+++ sdk/search/Azure.Search.Documents/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking Changes
 
 - Moved models for managing indexes, indexers, and skillsets to `Azure.Search.Documents.Indexes.Models`.
+- Split `SearchServiceClient` into `SearchIndexClient` for managing indexes, and `SearchIndexerClient` for managing indexers, both of which are now in `Azure.Search.Documents.Indexes`.
 
 ## 1.0.0-preview.3 (2020-05-05)
