+++ sdk/search/Azure.Search.Documents/tests/Samples/HelloWorld.cs
@@ -34,8 +34,8 @@ namespace Azure.Search.Documents.Tests.Samples
                 Environment.GetEnvironmentVariable("SEARCH_API_KEY"));
 
             // Create a new SearchServiceClient
-            SearchServiceClient search = new SearchServiceClient(endpoint, credential);
-            /*@@*/ search = InstrumentClient(new SearchServiceClient(endpoint, credential, GetSearchClientOptions()));
+            SearchIndexClient search = new SearchIndexClient(endpoint, credential);
+            /*@@*/ search = InstrumentClient(new SearchIndexClient(endpoint, credential, GetSearchClientOptions()));
 
             // Perform an operation
             Response<SearchServiceStatistics> stats = search.GetServiceStatistics();
@@ -60,8 +60,8 @@ namespace Azure.Search.Documents.Tests.Samples
                 Environment.GetEnvironmentVariable("SEARCH_API_KEY"));
 
             // Create a new SearchServiceClient
-            SearchServiceClient search = new SearchServiceClient(endpoint, credential);
-            /*@@*/ search = InstrumentClient(new SearchServiceClient(endpoint, credential, GetSearchClientOptions()));
+            SearchIndexClient search = new SearchIndexClient(endpoint, credential);
+            /*@@*/ search = InstrumentClient(new SearchIndexClient(endpoint, credential, GetSearchClientOptions()));
 
             // Perform an operation
             Response<SearchServiceStatistics> stats = await search.GetServiceStatisticsAsync();
@@ -139,8 +139,8 @@ namespace Azure.Search.Documents.Tests.Samples
             Uri endpoint = new Uri(Environment.GetEnvironmentVariable("SEARCH_ENDPOINT"));
             AzureKeyCredential credential = new AzureKeyCredential(
                 Environment.GetEnvironmentVariable("SEARCH_API_KEY"));
-            SearchServiceClient search = new SearchServiceClient(endpoint, credential);
-            /*@@*/ search = InstrumentClient(new SearchServiceClient(endpoint, credential, GetSearchClientOptions()));
+            SearchIndexClient search = new SearchIndexClient(endpoint, credential);
+            /*@@*/ search = InstrumentClient(new SearchIndexClient(endpoint, credential, GetSearchClientOptions()));
 
             // Get and report the Search Service statistics
             Response<SearchServiceStatistics> stats = await search.GetServiceStatisticsAsync();
