+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/ClientRequestIdRountripsAsync.json
