+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/Constructor.json
