+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/ConstructorAsync.json
