+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/CreateIndex.json
