+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/CreateIndexAsync.json
