+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/CreateIndexParameterValidation.json
