+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/CreateIndexParameterValidationAsync.json
