+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/CrudSynonymMaps.json
