+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/CrudSynonymMapsAsync.json
