+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/DiagnosticsAreUnique.json
