+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/DiagnosticsAreUniqueAsync.json
