+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/GetIndex.json
