+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/GetIndexAsync.json
