+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/GetIndexParameterValidation.json
