+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/GetIndexParameterValidationAsync.json
