+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/GetIndexes.json
