+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/GetSearchClient.json
