+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/GetSearchClientAsync.json
