+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/GetServiceStatistics.json
