+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/GetServiceStatisticsAsync.json
