+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/IndexSharesPipeline.json
