+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/IndexSharesPipelineAsync.json
