+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/SearchApiKeyCredential.json
