+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/SearchApiKeyCredentialAsync.json
