+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/UpdateIndex.json
