+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexClientTests/UpdateIndexAsync.json
