+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexerClientTests/ClientRequestIdRountrips.json
@@ -1,0 +1,46 @@
+{
+  "Entries": [
+    {
+      "RequestUri": "https://azs-net-heathsrchtst.search.windows.net/indexers?$select=%2A\u0026api-version=2019-05-06-Preview",
+      "RequestMethod": "GET",
+      "RequestHeaders": {
+        "Accept": "application/json; odata.metadata=minimal",
+        "api-key": "Sanitized",
+        "traceparent": "00-bdcdf5e6f7df4c40a641e3a76d252fac-d5dd5da48e18434b-00",
+        "User-Agent": [
+          "azsdk-net-Search.Documents/1.0.0-dev.20200519.1",
+          "(.NET Core 4.6.28801.04; Microsoft Windows 10.0.19041 )"
+        ],
+        "x-ms-client-request-id": "68966ae8-323d-1ca5-12f0-8e47cdae07d4",
+        "x-ms-return-client-request-id": "true"
+      },
+      "RequestBody": null,
+      "StatusCode": 200,
+      "ResponseHeaders": {
+        "Cache-Control": "no-cache",
+        "client-request-id": "68966ae8-323d-1ca5-12f0-8e47cdae07d4",
+        "Content-Length": "101",
+        "Content-Type": "application/json; odata.metadata=minimal",
+        "Date": "Tue, 19 May 2020 23:41:13 GMT",
+        "elapsed-time": "18",
+        "Expires": "-1",
+        "OData-Version": "4.0",
+        "Pragma": "no-cache",
+        "Preference-Applied": "odata.include-annotations=\u0022*\u0022",
+        "request-id": "68966ae8-323d-1ca5-12f0-8e47cdae07d4",
+        "Strict-Transport-Security": "max-age=15724800; includeSubDomains",
+        "x-ms-client-request-id": "68966ae8-323d-1ca5-12f0-8e47cdae07d4"
+      },
+      "ResponseBody": {
+        "@odata.context": "https://azs-net-heathsrchtst.search.windows.net/$metadata#indexers(*)",
+        "value": []
+      }
+    }
+  ],
+  "Variables": {
+    "RandomSeed": "1622410931",
+    "SearchIndexName": "iakmaxlf",
+    "SEARCH_ADMIN_API_KEY": "Sanitized",
+    "SEARCH_SERVICE_NAME": "azs-net-heathsrchtst"
+  }
+}
