+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexerClientTests/ClientRequestIdRountripsAsync.json
@@ -1,0 +1,46 @@
+{
+  "Entries": [
+    {
+      "RequestUri": "https://azs-net-heathsrchtst.search.windows.net/indexers?$select=%2A\u0026api-version=2019-05-06-Preview",
+      "RequestMethod": "GET",
+      "RequestHeaders": {
+        "Accept": "application/json; odata.metadata=minimal",
+        "api-key": "Sanitized",
+        "traceparent": "00-64a2166851960444b581875fa570defa-999e7943e4382642-00",
+        "User-Agent": [
+          "azsdk-net-Search.Documents/1.0.0-dev.20200519.1",
+          "(.NET Core 4.6.28801.04; Microsoft Windows 10.0.19041 )"
+        ],
+        "x-ms-client-request-id": "f32f3b11-7b2c-df0f-c417-672b2950b49a",
+        "x-ms-return-client-request-id": "true"
+      },
+      "RequestBody": null,
+      "StatusCode": 200,
+      "ResponseHeaders": {
+        "Cache-Control": "no-cache",
+        "client-request-id": "f32f3b11-7b2c-df0f-c417-672b2950b49a",
+        "Content-Length": "101",
+        "Content-Type": "application/json; odata.metadata=minimal",
+        "Date": "Tue, 19 May 2020 23:41:13 GMT",
+        "elapsed-time": "4",
+        "Expires": "-1",
+        "OData-Version": "4.0",
+        "Pragma": "no-cache",
+        "Preference-Applied": "odata.include-annotations=\u0022*\u0022",
+        "request-id": "f32f3b11-7b2c-df0f-c417-672b2950b49a",
+        "Strict-Transport-Security": "max-age=15724800; includeSubDomains",
+        "x-ms-client-request-id": "f32f3b11-7b2c-df0f-c417-672b2950b49a"
+      },
+      "ResponseBody": {
+        "@odata.context": "https://azs-net-heathsrchtst.search.windows.net/$metadata#indexers(*)",
+        "value": []
+      }
+    }
+  ],
+  "Variables": {
+    "RandomSeed": "1481307319",
+    "SearchIndexName": "iakmaxlf",
+    "SEARCH_ADMIN_API_KEY": "Sanitized",
+    "SEARCH_SERVICE_NAME": "azs-net-heathsrchtst"
+  }
+}
