+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexerClientTests/Constructor.json
@@ -1,0 +1,4 @@
+{
+  "Entries": [],
+  "Variables": {}
+}
