+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexerClientTests/CreateAzureBlobIndexer.json
