+++ sdk/search/Azure.Search.Documents/tests/SessionRecords/SearchIndexerClientTests/CreateAzureBlobIndexerAsync.json
